seat_recommendation/requirements.txt
# 座席推薦システムの依存パッケージ

# コアライブラリ
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# 設定ファイル
PyYAML>=5.4.0

# オプション: 可視化（データ探索用）
matplotlib>=3.4.0
seaborn>=0.11.0

# オプション: より高度なモデル
# xgboost>=1.5.0
# lightgbm>=3.3.0
